move platform code from 'server' package to 'platform' package
