+++ platform/src/main/java/org/stellar/anchor/platform/ConfigManagementConfig.java
@@ -1,4 +1,4 @@
-package org.stellar.anchor.server;
+package org.stellar.anchor.platform;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
